Refactor: Split JSNodeHTTPServerSocket into separate files
Moved JSNodeHTTPServerSocket and JSNodeHTTPServerSocketPrototype from
NodeHTTP.cpp into their own files following the same pattern as
JSDiffieHellman in the crypto module.

Changes:
- Created JSNodeHTTPServerSocket.h/cpp with class implementation
- Created JSNodeHTTPServerSocketPrototype.h/cpp with prototype methods
- Moved all custom getters/setters to prototype file
- Moved all host functions (close, write, end) to prototype file
- Preserved all extern C bindings for Zig interop
- Reduced NodeHTTP.cpp from ~1766 to 1010 lines

All tests pass (72/74 in node-http.test.ts, same as before).
