+++ src/bun.js/bindings/node/JSNodeHTTPServerSocketPrototype.h
@@ -1,0 +1,45 @@
+#pragma once
+
+#include "root.h"
+#include <JavaScriptCore/CallData.h>
+#include <JavaScriptCore/ObjectConstructor.h>
+
+namespace Bun {
+
+class JSNodeHTTPServerSocketPrototype final : public JSC::JSNonFinalObject {
+public:
+    using Base = JSC::JSNonFinalObject;
+    static constexpr unsigned StructureFlags = Base::StructureFlags | JSC::HasStaticPropertyTable;
+
+    static JSNodeHTTPServerSocketPrototype* create(JSC::VM& vm, JSC::Structure* structure)
+    {
+        JSNodeHTTPServerSocketPrototype* prototype = new (NotNull, JSC::allocateCell<JSNodeHTTPServerSocketPrototype>(vm)) JSNodeHTTPServerSocketPrototype(vm, structure);
+        prototype->finishCreation(vm);
+        return prototype;
+    }
+
+    template<typename, JSC::SubspaceAccess>
+    static JSC::GCClient::IsoSubspace* subspaceFor(JSC::VM& vm)
+    {
+        return &vm.plainObjectSpace();
+    }
+
+    DECLARE_INFO;
+
+    static JSC::Structure* createStructure(JSC::VM& vm, JSC::JSGlobalObject* globalObject, JSC::JSValue prototype)
+    {
+        auto* structure = JSC::Structure::create(vm, globalObject, prototype, JSC::TypeInfo(JSC::ObjectType, StructureFlags), info());
+        structure->setMayBePrototype(true);
+        return structure;
+    }
+
+private:
+    JSNodeHTTPServerSocketPrototype(JSC::VM& vm, JSC::Structure* structure)
+        : Base(vm, structure)
+    {
+    }
+
+    void finishCreation(JSC::VM&);
+};
+
+} // namespace Bun
